[item] rename default item struct
