+++ src/item.rs
@@ -1,8 +1,10 @@
 ///! An item is line of text that read from `find` command or stdin together with
 ///! the internal states, such as selected or not
-use crate::ansi::AnsiString;
+use crate::ansi::{ANSIParser, AnsiString};
+use crate::field::{parse_matching_fields, parse_transform_fields, FieldRange};
 use crate::spinlock::{SpinLock, SpinLockGuard};
 use crate::{ItemPreview, SkimItem};
+use regex::Regex;
 use std::borrow::Cow;
 use std::cmp::min;
 use std::default::Default;
@@ -10,6 +12,121 @@ use std::ops::Deref;
 use std::sync::atomic::{AtomicUsize, Ordering};
 use std::sync::Arc;
 
+//------------------------------------------------------------------------------
+/// An item will store everything that one line input will need to be operated and displayed.
+///
+/// What's special about an item?
+/// The simplest version of an item is a line of string, but things are getting more complex:
+/// - The conversion of lower/upper case is slow in rust, because it involds unicode.
+/// - We may need to interpret the ANSI codes in the text.
+/// - The text can be transformed and limited while searching.
+///
+/// About the ANSI, we made assumption that it is linewise, that means no ANSI codes will affect
+/// more than one line.
+#[derive(Debug)]
+pub struct DefaultSkimItem {
+    // The text that will be ouptut when user press `enter`
+    orig_text: String,
+
+    // The text that will shown into the screen. Can be transformed.
+    text: AnsiString<'static>,
+
+    matching_ranges: Vec<(usize, usize)>,
+
+    // For the transformed ANSI case, the output will need another transform.
+    using_transform_fields: bool,
+    ansi_enabled: bool,
+}
+
+impl<'a> DefaultSkimItem {
+    pub fn new(
+        orig_text: Cow<str>,
+        ansi_enabled: bool,
+        trans_fields: &[FieldRange],
+        matching_fields: &[FieldRange],
+        delimiter: &Regex,
+    ) -> Self {
+        let using_transform_fields = !trans_fields.is_empty();
+
+        //        transformed | ANSI             | output
+        //------------------------------------------------------
+        //                    +- T -> trans+ANSI | ANSI
+        //                    |                  |
+        //      +- T -> trans +- F -> trans      | orig
+        // orig |                                |
+        //      +- F -> orig  +- T -> ANSI     ==| ANSI
+        //                    |                  |
+        //                    +- F -> orig       | orig
+
+        let mut ansi_parser: ANSIParser = Default::default();
+
+        let text = if using_transform_fields && ansi_enabled {
+            // ansi and transform
+            ansi_parser.parse_ansi(&parse_transform_fields(delimiter, &orig_text, trans_fields))
+        } else if using_transform_fields {
+            // transformed, not ansi
+            AnsiString::new_string(parse_transform_fields(delimiter, &orig_text, trans_fields))
+        } else if ansi_enabled {
+            // not transformed, ansi
+            ansi_parser.parse_ansi(&orig_text)
+        } else {
+            // normal case
+            AnsiString::new_empty()
+        };
+
+        let mut ret = DefaultSkimItem {
+            orig_text: orig_text.into(),
+            text,
+            using_transform_fields: !trans_fields.is_empty(),
+            matching_ranges: Vec::new(),
+            ansi_enabled,
+        };
+
+        let matching_ranges = if !matching_fields.is_empty() {
+            parse_matching_fields(delimiter, &ret.get_text(), matching_fields)
+        } else {
+            vec![(0, ret.get_text().len())]
+        };
+
+        ret.matching_ranges = matching_ranges;
+        ret
+    }
+}
+
+impl SkimItem for DefaultSkimItem {
+    fn display(&self) -> Cow<AnsiString> {
+        if self.using_transform_fields || self.ansi_enabled {
+            Cow::Borrowed(&self.text)
+        } else {
+            Cow::Owned(AnsiString::new_str(&self.orig_text))
+        }
+    }
+
+    fn get_text(&self) -> Cow<str> {
+        if !self.using_transform_fields && !self.ansi_enabled {
+            Cow::Borrowed(&self.orig_text)
+        } else {
+            Cow::Borrowed(self.text.stripped())
+        }
+    }
+
+    fn output(&self) -> Cow<str> {
+        if self.using_transform_fields && self.ansi_enabled {
+            let mut ansi_parser: ANSIParser = Default::default();
+            let text = ansi_parser.parse_ansi(&self.orig_text);
+            Cow::Owned(text.into_inner())
+        } else if !self.using_transform_fields && self.ansi_enabled {
+            Cow::Borrowed(self.text.stripped())
+        } else {
+            Cow::Borrowed(&self.orig_text)
+        }
+    }
+
+    fn get_matching_ranges(&self) -> Cow<[(usize, usize)]> {
+        Cow::Borrowed(&self.matching_ranges)
+    }
+}
+//------------------------------------------------------------------------------
 pub struct ItemWrapper {
     // (num of run, number of index)
     id: (usize, usize),
@@ -36,10 +153,6 @@ impl SkimItem for ItemWrapper {
         self.inner.display()
     }
 
-    fn preview(&self) -> ItemPreview {
-        self.inner.preview()
-    }
-
     fn get_text(&self) -> Cow<str> {
         self.inner.get_text()
     }
@@ -48,11 +161,16 @@ impl SkimItem for ItemWrapper {
         self.inner.output()
     }
 
+    fn preview(&self) -> ItemPreview {
+        self.inner.preview()
+    }
+
     fn get_matching_ranges(&self) -> Cow<[(usize, usize)]> {
         self.inner.get_matching_ranges()
     }
 }
 
+//------------------------------------------------------------------------------
 #[derive(Debug, Copy, Clone, PartialEq, Default)]
 pub struct Rank {
     pub score: i64,
@@ -110,6 +228,7 @@ impl MatchedItem {
     }
 }
 
+//------------------------------------------------------------------------------
 const ITEM_POOL_CAPACITY: usize = 1024;
 
 pub struct ItemPool {
